coole notes app project completed

+++ backend/src/controllers/notesController.ts
@@ -2,11 +2,13 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export  const getNotes:RequestHandler = async(req,res,next)=>{
+    const authenticatedUserId = req.session.userId;
     try{
-      
-        const notes = await NoteModel.find().exec();      
+      assertIsDefined(authenticatedUserId);
+        const notes = await NoteModel.find({userId:authenticatedUserId}).exec();      
         res.status(200).json(notes);
     }catch(error){
        next(error);
@@ -16,13 +18,18 @@ export  const getNotes:RequestHandler = async(req,res,next)=>{
 
 export const getNote:RequestHandler = async(req,res,next)=>{
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,"Invalid Note Id")
         }
         const note = await NoteModel.findById(noteId).exec();
         if(!note){
          throw createHttpError(404,"Note Not Found")   
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You cannot access this note!!");
         }
         res.status(200).json(note)
     } catch (error) {
@@ -37,11 +44,15 @@ interface CreateNoteBody{
 export const createNote:RequestHandler<unknown,unknown,CreateNoteBody,unknown> = async(req,res,next)=>{
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
     try {
+        
+        assertIsDefined(authenticatedUserId);
         if(!title){
             throw createHttpError(400,"Note must have a title")
         }
         const newNote = await NoteModel.create({
+            userId:authenticatedUserId,
             title:title,
             text:text,
         })
@@ -62,7 +73,9 @@ export const updateNote:RequestHandler<updateNoteParams,unknown, updateNoteBody,
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,"Invalid Note Id")
         }
@@ -72,7 +85,10 @@ export const updateNote:RequestHandler<updateNoteParams,unknown, updateNoteBody,
         const note = await NoteModel.findById(noteId).exec();
         if(!note){
             throw createHttpError(404,"Note Not Found")   
-           }
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You cannot access this note!!");
+        }
            note.title = newTitle;
            note.text = newText;
            const updatedNote = await note.save();
@@ -83,10 +99,11 @@ export const updateNote:RequestHandler<updateNoteParams,unknown, updateNoteBody,
 }
 
 export const deleteNote:RequestHandler = async(req,res,next)=>{
-    console.log("dhukse");
     
+    const authenticatedUserId = req.session.userId;
     const noteId = req.params.noteId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400,"Invalid Note Id")
         }
@@ -94,7 +111,9 @@ export const deleteNote:RequestHandler = async(req,res,next)=>{
         if(!note){
             throw createHttpError(404,"Note not found");
         }
-        console.log("asdasdasd",note);        
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401,"You cannot access this note!!");
+        }   
         await note.deleteOne();
         res.sendStatus(204);
     } catch (error) {
